refactor: adicionar melhorias gerais no codigo

+++ mercado/src/main/java/com/example/mercado/entities/produto/ProdutoCreateDTO.java
@@ -1,4 +1,6 @@
 package com.example.mercado.entities.produto;
 
-public record ProdutoCreateDTO(String nome, Double valor, Long idMarca) {
+import java.math.BigDecimal;
+
+public record ProdutoCreateDTO(String nome, BigDecimal valor, Long idMarca) {
 }
